stopper: move to pkg package
This also adds a package docs for the stopper package.
